renamed directory to csharp. Moved algorithm to another file
